down to 7 circular dependencies
