+++ src/commands/node/handlers.ts
@@ -38,8 +38,8 @@ import {type PlatformInstaller} from '../../core/platform_installer.js';
 import {type K8} from '../../core/k8.js';
 import {type LeaseManager} from '../../core/lease/lease_manager.js';
 import {type RemoteConfigManager} from '../../core/config/remote/remote_config_manager.js';
-import {IllegalArgumentError} from '../../core/errors.js';
-import {ConsensusNodeStates} from '../../core/config/remote/enumerations.js';
+import {IllegalArgumentError, SoloError} from '../../core/errors.js';
+import {ComponentType, ConsensusNodeStates} from '../../core/config/remote/enumerations.js';
 import {RemoteConfigTasks} from '../../core/config/remote/remote_config_tasks.js';
 import type {SoloLogger} from '../../core/logging.js';
 import type {NodeCommand} from './index.js';
@@ -48,6 +48,12 @@ import {type Lease} from '../../core/lease/lease.js';
 import {NodeSubcommandType} from '../../core/enumerations.js';
 import {type CommandHandlers} from '../base.js';
 import {NodeHelper} from './helper.js';
+import type {NodeAlias, NodeAliases} from '../../types/aliases.js';
+import {ConsensusNodeComponent} from '../../core/config/remote/components/consensus_node_component.js';
+import {type Listr, type ListrTask} from 'listr2';
+import chalk from 'chalk';
+import type {ComponentsDataWrapper} from '../../core/config/remote/components_data_wrapper.js';
+import type {Optional} from '../../types/index.js';
 
 export class NodeCommandHandlers implements CommandHandlers {
   private readonly accountManager: AccountManager;
@@ -109,7 +115,7 @@ export class NodeCommandHandlers implements CommandHandlers {
     return [
       this.tasks.initialize(argv, deleteConfigBuilder.bind(this), lease),
       RemoteConfigTasks.loadRemoteConfig.bind(this)(argv),
-      RemoteConfigTasks.validateSingleNodeState.bind(this)({excludedStates: []}),
+      NodeCommandHandlers.validateSingleNodeState.bind(this)({excludedStates: []}),
       this.tasks.identifyExistingNodes(),
       this.tasks.loadAdminKey(),
       this.tasks.prepareUpgradeZip(),
@@ -153,7 +159,7 @@ export class NodeCommandHandlers implements CommandHandlers {
     return [
       this.tasks.initialize(argv, addConfigBuilder.bind(this), lease),
       RemoteConfigTasks.loadRemoteConfig.bind(this)(argv),
-      RemoteConfigTasks.validateSingleNodeState.bind(this)({excludedStates: []}),
+      NodeCommandHandlers.validateSingleNodeState.bind(this)({excludedStates: []}),
       this.tasks.checkPVCsEnabled(),
       this.tasks.identifyExistingNodes(),
       this.tasks.determineNewNodeAccountNumber(),
@@ -206,7 +212,7 @@ export class NodeCommandHandlers implements CommandHandlers {
     return [
       this.tasks.initialize(argv, updateConfigBuilder.bind(this), lease),
       RemoteConfigTasks.loadRemoteConfig.bind(this)(argv),
-      RemoteConfigTasks.validateSingleNodeState.bind(this)({excludedStates: []}),
+      NodeCommandHandlers.validateSingleNodeState.bind(this)({excludedStates: []}),
       this.tasks.identifyExistingNodes(),
       this.tasks.loadAdminKey(),
       this.tasks.prepareUpgradeZip(),
@@ -632,7 +638,7 @@ export class NodeCommandHandlers implements CommandHandlers {
       [
         this.tasks.initialize(argv, refreshConfigBuilder.bind(this), lease),
         RemoteConfigTasks.loadRemoteConfig.bind(this)(argv),
-        RemoteConfigTasks.validateAllNodeStates.bind(this)({
+        NodeCommandHandlers.validateAllNodeStates.bind(this)({
           acceptedStates: [ConsensusNodeStates.STARTED, ConsensusNodeStates.SETUP, ConsensusNodeStates.INITIALIZED],
         }),
         this.tasks.identifyNetworkPods(),
@@ -687,12 +693,12 @@ export class NodeCommandHandlers implements CommandHandlers {
       [
         this.tasks.initialize(argv, stopConfigBuilder.bind(this), lease),
         RemoteConfigTasks.loadRemoteConfig.bind(this)(argv),
-        RemoteConfigTasks.validateAllNodeStates.bind(this)({
+        NodeCommandHandlers.validateAllNodeStates.bind(this)({
           acceptedStates: [ConsensusNodeStates.STARTED, ConsensusNodeStates.SETUP],
         }),
         this.tasks.identifyNetworkPods(),
         this.tasks.stopNodes(),
-        RemoteConfigTasks.changeAllNodeStates.bind(this)(ConsensusNodeStates.INITIALIZED),
+        NodeCommandHandlers.changeAllNodeStates.bind(this)(ConsensusNodeStates.INITIALIZED),
       ],
       {
         concurrent: false,
@@ -715,14 +721,14 @@ export class NodeCommandHandlers implements CommandHandlers {
       [
         this.tasks.initialize(argv, startConfigBuilder.bind(this), lease),
         RemoteConfigTasks.loadRemoteConfig.bind(this)(argv),
-        RemoteConfigTasks.validateAllNodeStates.bind(this)({acceptedStates: [ConsensusNodeStates.SETUP]}),
+        NodeCommandHandlers.validateAllNodeStates.bind(this)({acceptedStates: [ConsensusNodeStates.SETUP]}),
         this.tasks.identifyExistingNodes(),
         this.tasks.uploadStateFiles((ctx: any) => ctx.config.stateFile.length === 0),
         this.tasks.startNodes('nodeAliases'),
         this.tasks.enablePortForwarding(),
         this.tasks.checkAllNodesAreActive('nodeAliases'),
         this.tasks.checkNodeProxiesAreActive(),
-        RemoteConfigTasks.changeAllNodeStates.bind(this)(ConsensusNodeStates.STARTED),
+        NodeCommandHandlers.changeAllNodeStates.bind(this)(ConsensusNodeStates.STARTED),
         this.tasks.addNodeStakes(),
       ],
       {
@@ -746,13 +752,13 @@ export class NodeCommandHandlers implements CommandHandlers {
       [
         this.tasks.initialize(argv, setupConfigBuilder.bind(this), lease),
         RemoteConfigTasks.loadRemoteConfig.bind(this)(argv),
-        RemoteConfigTasks.validateAllNodeStates.bind(this)({
+        NodeCommandHandlers.validateAllNodeStates.bind(this)({
           acceptedStates: [ConsensusNodeStates.INITIALIZED],
         }),
         this.tasks.identifyNetworkPods(),
         this.tasks.fetchPlatformSoftware('nodeAliases'),
         this.tasks.setupNetworkNodes('nodeAliases'),
-        RemoteConfigTasks.changeAllNodeStates.bind(this)(ConsensusNodeStates.SETUP),
+        NodeCommandHandlers.changeAllNodeStates.bind(this)(ConsensusNodeStates.SETUP),
       ],
       {
         concurrent: false,
@@ -764,5 +770,154 @@ export class NodeCommandHandlers implements CommandHandlers {
 
     await action(argv, this);
     return true;
+  }
+
+  /** Removes the consensus node, envoy and haproxy components from remote config.  */
+  public static removeNodeAndProxies(this: NodeCommandHandlers): ListrTask<any, any, any> {
+    return {
+      skip: (): boolean => !this.remoteConfigManager.isLoaded(),
+      title: 'Remove node and proxies from remote config',
+      task: async (): Promise<void> => {
+        await this.remoteConfigManager.modify(async remoteConfig => {
+          remoteConfig.components.remove('Consensus node name', ComponentType.ConsensusNode);
+          remoteConfig.components.remove('Envoy proxy name', ComponentType.EnvoyProxy);
+          remoteConfig.components.remove('HaProxy name', ComponentType.HaProxy);
+        });
+      },
+    };
+  }
+
+  /**
+   * Changes the state from all consensus nodes components in remote config.
+   *
+   * @param state - to which to change the consensus node component
+   */
+  public static changeAllNodeStates(this: NodeCommandHandlers, state: ConsensusNodeStates): ListrTask<any, any, any> {
+    interface Context {
+      config: {namespace: string; nodeAliases: NodeAliases};
+    }
+
+    return {
+      title: `Change node state to ${state} in remote config`,
+      skip: (): boolean => !this.remoteConfigManager.isLoaded(),
+      task: async (ctx: Context): Promise<void> => {
+        await this.remoteConfigManager.modify(async remoteConfig => {
+          const {
+            config: {namespace, nodeAliases},
+          } = ctx;
+          const cluster = this.remoteConfigManager.currentCluster;
+
+          for (const nodeAlias of nodeAliases) {
+            remoteConfig.components.edit(nodeAlias, new ConsensusNodeComponent(nodeAlias, cluster, namespace, state));
+          }
+        });
+      },
+    };
+  }
+
+  /**
+   * Creates tasks to validate that each node state is either one of the accepted states or not one of the excluded.
+   *
+   * @param acceptedStates - the state at which the nodes can be, not matching any of the states throws an error
+   * @param excludedStates - the state at which the nodes can't be, matching any of the states throws an error
+   */
+  public static validateAllNodeStates(
+    this: NodeCommandHandlers,
+    {acceptedStates, excludedStates}: {acceptedStates?: ConsensusNodeStates[]; excludedStates?: ConsensusNodeStates[]},
+  ): ListrTask<any, any, any> {
+    interface Context {
+      config: {namespace: string; nodeAliases: NodeAliases};
+    }
+
+    return {
+      title: 'Validate nodes states',
+      skip: (): boolean => !this.remoteConfigManager.isLoaded(),
+      task: (ctx: Context, task): Listr<any, any, any> => {
+        const nodeAliases = ctx.config.nodeAliases;
+
+        const components = this.remoteConfigManager.components;
+
+        const subTasks: ListrTask<Context, any, any>[] = nodeAliases.map(nodeAlias => ({
+          title: `Validating state for node ${nodeAlias}`,
+          task: (_, task): void => {
+            const state = NodeCommandHandlers.validateNodeState(nodeAlias, components, acceptedStates, excludedStates);
+
+            task.title += ` - ${chalk.green('valid state')}: ${chalk.cyan(state)}`;
+          },
+        }));
+
+        return task.newListr(subTasks, {
+          concurrent: false,
+          rendererOptions: {collapseSubtasks: false},
+        });
+      },
+    };
+  }
+
+  /**
+   * Creates tasks to validate that specific node state is either one of the accepted states or not one of the excluded.
+   *
+   * @param acceptedStates - the state at which the node can be, not matching any of the states throws an error
+   * @param excludedStates - the state at which the node can't be, matching any of the states throws an error
+   */
+  public static validateSingleNodeState(
+    this: NodeCommandHandlers,
+    {acceptedStates, excludedStates}: {acceptedStates?: ConsensusNodeStates[]; excludedStates?: ConsensusNodeStates[]},
+  ): ListrTask<any, any, any> {
+    interface Context {
+      config: {namespace: string; nodeAlias: NodeAlias};
+    }
+
+    return {
+      title: 'Validate nodes state',
+      skip: (): boolean => !this.remoteConfigManager.isLoaded(),
+      task: (ctx: Context, task): void => {
+        const nodeAlias = ctx.config.nodeAlias;
+
+        task.title += ` ${nodeAlias}`;
+
+        const components = this.remoteConfigManager.components;
+
+        const state = NodeCommandHandlers.validateNodeState(nodeAlias, components, acceptedStates, excludedStates);
+
+        task.title += ` - ${chalk.green('valid state')}: ${chalk.cyan(state)}`;
+      },
+    };
+  }
+
+  /**
+   * @param nodeAlias - the alias of the node whose state to validate
+   * @param components - the component data wrapper
+   * @param acceptedStates - the state at which the node can be, not matching any of the states throws an error
+   * @param excludedStates - the state at which the node can't be, matching any of the states throws an error
+   */
+  private static validateNodeState(
+    nodeAlias: NodeAlias,
+    components: ComponentsDataWrapper,
+    acceptedStates: Optional<ConsensusNodeStates[]>,
+    excludedStates: Optional<ConsensusNodeStates[]>,
+  ): ConsensusNodeStates {
+    let nodeComponent: ConsensusNodeComponent;
+    try {
+      nodeComponent = components.getComponent<ConsensusNodeComponent>(ComponentType.ConsensusNode, nodeAlias);
+    } catch (e) {
+      throw new SoloError(`${nodeAlias} not found in remote config`);
+    }
+
+    if (acceptedStates && !acceptedStates.includes(nodeComponent.state)) {
+      const errorMessageData =
+        `accepted states: ${acceptedStates.join(', ')}, ` + `current state: ${nodeComponent.state}`;
+
+      throw new SoloError(`${nodeAlias} has invalid state - ` + errorMessageData);
+    }
+
+    if (excludedStates && excludedStates.includes(nodeComponent.state)) {
+      const errorMessageData =
+        `excluded states: ${excludedStates.join(', ')}, ` + `current state: ${nodeComponent.state}`;
+
+      throw new SoloError(`${nodeAlias} has invalid state - ` + errorMessageData);
+    }
+
+    return nodeComponent.state;
   }
 }
